+++ packages/apollo-engine-reporting-protobuf/package.json
@@ -5,9 +5,9 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "prepare": "npm run pbjs && npm run pbts",
-    "pbjs": "bash -c 'mkdir -p dist && pbjs --target static-module --out dist/index.js --wrap commonjs --force-number <(grep -v \"package mdg.engine.proto\" reports.proto)'",
-    "pbts": "pbts -o dist/index.d.ts dist/index.js"
+    "prepare": "npm run pbjs && npm run pbts && cp src/* dist",
+    "pbjs": "bash -c 'mkdir -p dist && pbjs --target static-module --out dist/protobuf.js --wrap commonjs --force-number <(grep -v \"package mdg.engine.proto\" src/reports.proto)'",
+    "pbts": "pbts -o dist/protobuf.d.ts dist/protobuf.js"
   },
   "repository": {
     "type": "git",
