+++ packages/apollo-engine-reporting-protobuf/src/index.d.ts
@@ -1,0 +1,2 @@
+import * as protobuf from './protobuf';
+export = protobuf;
